Remove old repeat mod
