+++ hitori-macros/src/expand/matches_block.rs
@@ -1,5 +1,4 @@
 mod repeat;
-mod repeat_new;
 mod state;
 
 use crate::utils::{
